datagen/requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Faker==1.0.7
google-api-core==1.11.1
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-bigquery==1.14.0
google-cloud-core==1.0.1
google-cloud-pubsub==0.41.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.21.1
idna==2.8
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
text-unidecode==1.2
urllib3==1.25.3
